chore: move OSR related things in SetOwnerWindow to CommonWebContentsDelegate
